Use binary data packets instead of JSON to save transfer bandwidth

+++ CarbonMonoxide.hpp
@@ -1,9 +1,11 @@
 #ifndef CARBON_MONOXIDE_HPP
 #define CARBON_MONOXIDE_HPP
 
+#include <stdint.h>
+
 class CarbonMonoxide {
 public:
-	int value;
+	int16_t value;
 
 	CarbonMonoxide();
 	void init();
